Configure Spring Boot and Postgres in application
The Spring Boot and the Postgres JDBC driver are now added in the build.gradle file as dependencies. The application.properties file is also created to establish connection with the Postgres database through spring.datasource properties. The hibernate ddl-auto is set to none and SQL statements are shown in console output during runtime.