Remove user blocking in UserService and refactor PostService and Logic
Removed the functionality for setting user block status from UserService. Instead, the block status is now directly set in the Logic class. The parameters order in the 'createAndSavePost' method in PostService was also altered. Additionally, comprehensive user request handling logic related to post creation was added to the Logic class.