Replace log4j2 XML configuration with properties file
This change removes the XML configuration for log4j2 and replaces it with a properties file. The properties based configuration provides the same logging capabilities as the XML but in a format that is often easier to read and manage. Also, an additional appender for file output has been included.