Update Docker and Docker-compose settings
This update refines settings in docker-compose.yaml and Dockerfile. The docker-compose.yaml now sets the app to restart unless stopped, and adds a dedicated container name and volume for logs. The Dockerfile now specifies a location for the user's Gradle home. These modifications will lead to more consistent service continuity and better management of logs and builds.